ai-automation-workflows-llm/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# LLM providers
openai==1.3.7
anthropic==0.7.8

# Data processing and analysis
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0

# Deep Learning
tensorflow==2.15.0
keras==2.15.0
torch==2.1.1
torchvision==0.16.1

# Computer Vision and Gesture Recognition
opencv-python==4.13.0.92
mediapipe==0.10.8
Pillow==10.1.0

# Voice Recognition and Text-to-Speech
SpeechRecognition==3.10.1
pyttsx3==2.90
pyaudio==0.2.14

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
bokeh==3.2.2

# AIML (Artificial Intelligence Markup Language)
aiml==0.9.3
nltk==3.8.1

# HTTP and async
httpx==0.25.2
aiofiles==23.2.1

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1
click==8.1.7

# Logging and monitoring
structlog==23.2.0
wandb==0.16.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Jupyter and notebooks
jupyter==1.0.0
ipykernel==6.26.0
notebook==7.0.6

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# Database and storage
sqlalchemy==2.0.23
alembic==1.13.0
redis==5.0.1
celery==5.3.4

# Image processing
imageio==2.31.6
scikit-image==0.22.0

# Additional ML tools
imbalanced-learn==0.11.0
shap==0.42.1
eli5==0.13.0

# Web scraping and data collection
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2

# Natural Language Processing
spacy==3.7.2
textblob==0.17.1
gensim==4.3.2

# Time series analysis
statsmodels==0.14.0
prophet==1.1.4

# Geospatial analysis
geopandas==0.14.1
folium==0.15.0

# Performance optimization
numba==0.58.1
cython==3.0.8

# Security and encryption
cryptography==41.0.7
passlib==1.7.4

# API and web tools
flask==3.0.0
django==4.2.7
streamlit==1.28.1

# Cloud integration
boto3==1.29.7
google-cloud-storage==2.10.0
azure-storage-blob==12.19.0
